ref: upstream lister and it's impl to module pkg
